Verify email is in claims before creating user
Previously it could occur that during authentication a user would try to be created without the email claim. However the create_user function would now allow for this, returning `None`. However on the surface everything seems to work fine, with no debug or warning messages to speak of. Instead of always returning `True` for `verify_claims` it would therefor make more sense if this function actually checked if the email is present before trying to create the user.

If this would break existing functionality perhaps some explicit warning message should be logged when the `create_user` function would not create the user and return `None` to prevent hours of debugging for people :)