Add OIDCClient for django tests
Add an OIDCClient for django tests. This can be useful when you need to use the oidc_id_token in your views.
The default Django test client does not provide this. By making use of this client your requests will contain
a `oidc_id_token` key with a random value (which you can specify). Because there is no easy way to do
this yourself for the Django test client (you can either fix it by writing a custom client or by using
a `RequestFactory`, both are not very ideal) I thought it would be useful to provide a default client
that comes with the mozilla library.