Fix JWKS handling with same `kid` across JWKs
Prior to this change, when a JWKS was delivered with multiple JWK with
the same `kid` but different `alg` specified, the function would
consider the first key in the list where the `kid` matched.  If the
`alg` didn't match, an exception would be raised and if there was a
valid JWK with the right `alg` (eg later in the list) it would not be
considered.

In the JWK spec (https://tools.ietf.org/html/rfc7517#section-4.4), the
`alg` value isn't required to be unique -- only to match -- so the check
shouldn't raise an exception at this point, but rather continue checking
all remaining JWKs.

With this change, an exception is still raised at the end if no valid
JWK is found.

Refines the change at https://github.com/mozilla/mozilla-django-oidc/commit/decc46f572a59f17bdcb2eade8a1ba37ee1fe7d8