Refs #348 Only call auth.login() when the user has changed
Adds login to prevent calling the auth.login function when the user hasn't
actually changed. This is the case when the session is being just being
refreshed, versus a real full login. This prevents cycling the CSRF token for
the session, which thereby fixes the invalid CSRF token errors described in
bug #348.