Verify user_can_authenticate when authenticating user
The standard Django `ModelBackend` only returns a user from  `authenticate`
if `user_can_authenticate` is `True`.

The `OIDCAuthenticationBackend` doesn't check `user_can_authenticate`,
in its `authenticate` method, allowing a user to log-in that has `is_active`
set to `False`.

This behavior is a surprising deviation from the way `ModelBackend`
works.

It can lead to security problems because other important
authentication checks may inadvertently be bypassed, if the user
implements them in `user_can_authenticate`, expecting
`OIDCAuthenticationBackend` to call it the same as any other
`ModelBackend`.

It seems proper for the `OIDCAuthenticationBackend` to verify
`user_can_authenticate` and it seems to me like there are no problems
associated with doing so, so let's just call `user_can_authenticate`
to resolve the problems described above.