Track login next URL as part of OIDC state
We support multiple parallel login processes by using a dictionary
for `oidc_state`.

With only a single shared `oidc_login_next` field in the session,
all login processes will redirect the user to the next URL set by
the login process they initiated last.

This behavior seems confusing and not that useful.

By storing the login next URL as part of the state, we can have each
login process redirect to the next URL provided at its initiation.

Also, we get the benefit that the login next URL for that login process
will not be lingering around in the session anymore, as the whole state
for that login process, including the login next URL, will be removed by
the OIDC callback.