Added on-demand session token refresh
Added a special case that allows AJAX queries to refresh the session
on demand.  AJAX queries simply need to add the
header 'X-Refresh-OIDC-Token' to the request and it will reset the
session expiration so it generates a reauth redirect.
This allows SPAs and other Javascript driven applications
to proactively control session refresh.